youtube-realtime-pipeline/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
python-dotenv
pydantic
google-api-python-client
streamlit
google-generativeai
dnspython
google-adk[web]